updated lint settings and removed lint warnings
