+++ .eslintignore
@@ -5,3 +5,6 @@ node_modules
 /out/
 
 /build
+
+# don't lint build output (make sure it's set to your correct build folder name)
+dist
