+++ package.json
@@ -14,6 +14,7 @@
     "aos": "^2.3.4",
     "autoprefixer": "^10.4.7",
     "eip-712": "^1.0.0",
+    "eslint-config-prettier": "^8.5.0",
     "ethers": "^5.6.8",
     "next": "12.1.6",
     "react": "18.1.0",
